done: no-admin user in wordpress db
learned: when building containers, remove files inside folders
that will be bind mounted in containers.

learned: in host, always delete all files that will be bind mounted
in containers.

inside host, all is working fine, but in vm, when we will create an
blog post, we still get the error message.